CMPE 246 IoT Project Code/requirements.txt
RPi.GPIO>=0.7,<1.0
spidev>=3.6,<4.0
requests>=2.31,<3.0
